+++ src/Util.hpp
@@ -5,43 +5,43 @@
 #include <iostream>
 
 [[nodiscard]]
-inline uint32_t getOpcode(uint32_t instruction) noexcept
+inline uint32_t getOpcode(uint64_t instruction) noexcept
 {
     return instruction & 0xF;
 }
 
 [[nodiscard]]
-inline uint32_t getSubop(uint32_t instruction) noexcept
+inline uint32_t getSubop(uint64_t instruction) noexcept
 {
     return (instruction >> 4) & 0x3;
 }
 
 [[nodiscard]]
-inline uint32_t getFpuop(uint32_t instruction) noexcept
+inline uint32_t getFpuop(uint64_t instruction) noexcept
 {
     return (instruction >> 4) & 0xF;
 }
 
 [[nodiscard]]
-inline uint32_t getRd(uint32_t instruction) noexcept
+inline uint32_t getRd(uint64_t instruction) noexcept
 {
     return (instruction >> 14) & 0x3F;
 }
 
 [[nodiscard]]
-inline uint32_t getRs1(uint32_t instruction) noexcept
+inline uint32_t getRs1(uint64_t instruction) noexcept
 {
     return (instruction >> 20) & 0x3F;
 }
 
 [[nodiscard]]
-inline uint32_t getRs2(uint32_t instruction) noexcept
+inline uint32_t getRs2(uint64_t instruction) noexcept
 {
     return (instruction >> 26) & 0x3F;
 }
 
 [[nodiscard]]
-inline int32_t getImmediate(uint32_t instruction) noexcept
+inline int32_t getImmediate(uint64_t instruction) noexcept
 {
     return (instruction >> 6) & 0x3FFF ;
 }
